Encrypt and decrypt values in service layer
Encrypt and decrypt values in KVService.

Removed getKeysWithValue(). With encryption, it is simply not feasable
to have that method since it is not possible to re-encrypt an incoming
value and have it produce the same cipher text as it previously did (
because of the IV used to encrypt the original text).

Unit tests work.