Finish implementing key-value signing
Fixed KVCrypt's sign() method to take in the key, in addition to the
value. It was previously only accepting the value. Then enhanced the
hmac routine to incorporate the key into the signature.

New method on KVCrypt to confirm signatures, confirmSignature().

Baked the signing processing right into KVDao. Even if encryption were
to be made optional, I believe that confirming the authenticity of a
record is essential. By implementing signing directly in the DAO, it
becomes more of a core feature.