Refactor crypt code
Moved all the key code to separate classes. There are now dedicated
classes for reading, writing, generating and managing keys.

Write all the keys to a single JSON file instead of separate files.

KeysReader and Writer are obvious.

Keys is a pseudo-model class for the keys. Its functionality is limited
to setting and getting keys. It can be locked from modifications and
its data can be formatted to JSON.

KeyStore is a manager class for Keys. It is used for initializing the
keys and retrieving them. It also locks its keys after initializion.

Key-value signing was also moved to a separate class: KVSigner.

KVCrypt is still a singleton for encrypting, decrypting and signing,
however, it is also a wrapper around an instance of KeyStore. Its
contents are much slimmer since all the key and signing code has been
abstracted.