Implement test mode
Redesigned config module to support a production mode and a test mode.

The object returned by the module is a wrapper object around the
currently active config (production or test). By default, get() will
return the production config, but calling enableTestMode() will
switch it over to the test config.

This cleans up the handling of test variables / configurations. One
thing that really annoyed about the previous setup waas that some of
the main modules had to check the value of config.testMode. Code simply
should not have to do that type of check.