Create new configuration file and loader
A new simpler JSON file config.js to store app configs. Along with this
is a loader class, ConfigLoader, and a unit test for it.

For now, the original tests have been removed. They will be returned
as parts of the project are cleaned up.

One of the goals of this new config system is to make educated decisions
and stick to them, rather than attempting to give the user a number of
confusing options.