Create script for building keys
Made script for creating keys and setup Makefile to use it.

Also updated KeyStore to have its own confirm() method that runs assert
statements on itself. Basically, I decided that I want the key store to
contain its own ability to validate itself, that way it can be used both
in the mocha test and as part of initialization procedures.