Add signing back into new KVCrypt
The logic remains mostly unchanged from the previous version. The
biggest difference is that a new model class, EncryptedKeyValue, was
created for the purpose of passing around a key along with its
corresponding encrypted IVCipher object. When confirming a signature,
rather than passing in an expected signature, two EncryptedKeyValue
objects are passesd in and then both a signed and checked. The benefit to
this approach is that the signature has to passed around code much less
than it previously had to be. For example, after values are pulled out
of the database for confirmation, they can simply be stuffed into their
model and then checked. Previously, they would have had to be first
signed before checked. This essentially keeps signing logic within
the confines of KVCrypt.

A mocha test was also added to confirm that all of this works.