Start implementing Mocha test for KVDao
Since testing runs in-memory, I had to implement an initialization
routine that would detect if the database was in-memory, and if it is,
then run a create table statement. The code for was puilled out of
KVTestHelper and wrapped up by a new Initializer class. KVTestHelper
still exists but just calls Initializer.