Reintergrate express with the application
Setting up and launching an express instance was integrated with the
component system. As part of this, I took the path of developing
dedicated Component subclasses for each component. The main reason for
doing this was so that each component can assume control over how it
should be uninitializaed/destroyed. This was important for express so
that the server can be cleanly shutdown while keeping those details out
of the component loader itself. To suppor this, the ComponentStore now has a destroy()
method that will destroy all of its components and then uninitialize
itself. This new functionality has also been integrated into all the unit
tests, such that mocha tests now have an after() that tearsdown the
component store.

As part of these changes, I also pulled out the use of the external async
module and replaced it with ECMA6 Promise.all().